+++ cart/shopping-cart.js
@@ -1,7 +1,7 @@
 import apparel from '../data/apparel.js';
 import cartData from '../data/cart.js';
 import renderLineItems from '../cart/render-line-items.js';
-import findById from '../common/utils.js';
+import findById, { calcOrderItem } from '../common/utils.js';
 
 
 const table = document.getElementById('table');
@@ -13,9 +13,14 @@ for (let i = 0; i < apparel.length; i ++) {
 
     const generateCart = renderLineItems(cartData, findCartItemsById);
 
+
+    
+
     table.appendChild(generateCart);
 }
 
 
+
+// Use the function to calculate the cart total and display the result in the appropriate element.
