+++ test/tests.js
@@ -1,1 +1,1 @@
-import './test.utils.js';
+import './utils.test.js';
